QT abi fixup

+++ src/3rdparty/webkit/WebKit/qt/Api/qwebsettings.h
@@ -102,11 +102,11 @@ public:
     void setDefaultTextEncoding(const QString &encoding);
     QString defaultTextEncoding() const;
 
-    void setPrintingMinimumShrinkFactor(float printingMinimumShrinkFactor);
-    float printingMinimumShrinkFactor() const;
+    void setPrintingMinimumShrinkFactor(qreal printingMinimumShrinkFactor);
+    qreal printingMinimumShrinkFactor() const;
 
-    void setPrintingMaximumShrinkFactor(float printingMaximimShrinkFactor);
-    float printingMaximumShrinkFactor() const;
+    void setPrintingMaximumShrinkFactor(qreal printingMaximimShrinkFactor);
+    qreal printingMaximumShrinkFactor() const;
 
     static void setIconDatabasePath(const QString &location);
     static QString iconDatabasePath();
